feat: convert cut to mp3
