+++ .gitignore
@@ -1,2 +1,3 @@
 /node_modules
 *.mp4
+*.mp3
