+++ src/index.js
@@ -1,6 +1,6 @@
 import { getDuration, askOptions, formatFileName, openFile } from './helpers'
 import { downloadFromYoutube, getVideoTitle } from './download'
-import { cutVideo } from './cut'
+import { cutVideo, convertToMp3 } from './convert'
 
 /**
  * Runs this package
@@ -8,12 +8,14 @@ import { cutVideo } from './cut'
 ;(async function main() {
   try {
     const answers = await askOptions()
-    const { openAtFinish, youtubeUrl, startTime, endTime } = answers
+    const { youtubeUrl, startTime, endTime, openAtFinish, toMp3 } = answers
 
     const videoTitle = await getVideoTitle(youtubeUrl)
     const fileName = formatFileName(videoTitle)
-    const downloadPath = `videos/full_${fileName}`
-    const convertPath = `videos/${fileName}`
+
+    const downloadPath = `resources/downloads/${fileName}.mp4`
+    const convertPath = `resources/cuts/${fileName}.mp4`
+    const audioPath = `resources/audios/${fileName}.mp3`
 
     const hasDownloaded = await downloadFromYoutube(youtubeUrl, downloadPath)
     if (!hasDownloaded) {
@@ -25,8 +27,14 @@ import { cutVideo } from './cut'
       throw new Error('>> Video was not cut')
     }
 
+    let hasConvertedToMp3 = false
+    if (toMp3) {
+      hasConvertedToMp3 = await convertToMp3(convertPath, audioPath)
+    }
+
     if (openAtFinish) {
-      openFile(convertPath)
+      const path = hasConvertedToMp3 ? audioPath : convertPath
+      openFile(path)
     }
   } catch (error) {
     console.error(error)
